+++ src/mongo/db/matcher/expression_internal.h
@@ -32,7 +32,7 @@ namespace mongo {
     // Replaces getFieldDottedOrArray without recursion nor string manipulation
     BSONElement getFieldDottedOrArray( const BSONObj& doc,
                                        const FieldRef& path,
-                                       int32_t* idxPath,
+                                       size_t* idxPath,
                                        bool* inArray );
 
 }  // namespace mongo
